# Faster babel