--- a/client/webpack.config.js
+++ b/client/webpack.config.js
@@ -11,6 +11,7 @@ module.exports = {
         loaders: [
             {
                 test: /\.jsx?$/,
+                exclude: /node_modules/,
                 loader: 'babel-loader',
                 query: {
                     presets: ['es2015']