--- a/client/webpack.config.js
+++ b/client/webpack.config.js
@@ -14,7 +14,8 @@ module.exports = {
                 exclude: /node_modules/,
                 loader: 'babel-loader',
                 query: {
-                    presets: ['es2015']
+                    presets: ['es2015'],
+                    cacheDirectory: true
                 }
             }
         ]